Refactored update_status and added thread_joining.
Update_status function now stops printing after the first philosopher dies and kills the rest of the philosophers so that their threads return.
Thread_joining has been implemented, so that the main thread waits for
the exiting of the other threads before returning. This should be
rechecked though, not sure if it was implmented correctly.